Added unordered list tag to list in file
Added unordered list tag to list in file to comply with W3C validation.
This is the error it was reporting and now fixed after this change was
made.

Line 384, Column 28: Element li not allowed as child of element div in
this context. (Suppressing further errors from this subtree.)

<li class="topicoption-off">Home </li>

Contexts in which element li may be used:
Inside ol elements.
Inside ul elements.
Inside menu elements whose type attribute is in the toolbar state.
Content model for element div:
Flow content.