Adds Character Counter To Fields
Mark I'm not sure how to add the proper javascript to switch in uk-badge-danger when character limit is exceeded but the javascript library does allow for an exceeded trigger. 

https://github.com/dtisgodsson/jquery-character-counter

Here on the options

{
limit: 150,
counterWrapper: 'span',
counterCssClass: 'counter',
counterFormat: '%1',
counterExceededCssClass: 'exceeded',
onExceed: function(count){},
onDeceed: function(count){},
customFields: {}
}

limit - the number of characters you wish to limit.
counterWrapper - the element you wish to wrap your counter in.
counterCssClass - the CSS class to apply to your counter.
counterFormat - the format of your counter text where '%1' will be replaced with the remaining character count.
counterExceededCssClass - the CSS class to apply when your limit has been exceeded.
onExceed - this function is called when the limit is breached
onDeceed - this function is called when the limit, having previously been exceeded, is now deceeded customFields - key value pairs of custom options to be added to the counter such as class, data attributes etc.