Change order of css files
The custom.css should be the last in the style.cache.css
I think it's better if you will make a new theme, so you can overwrite
all styles - also the plugin styles - with only 1 css file. At the
moment the plugin styles the last in the style.cache.css.

Maybe there is a ohther way but i think this is the easiest